// @types/graphql-types.d.ts
export type Date = any;

/** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
export type JSON = any;
/** An object with an id, parent, and children */
export interface Node {
  id: string; /** The id of the node. */
  parent?: Node | null; /** The parent of this node. */
  children?: (Node | null)[] | null; /** The children of this node. */
}
// ...
export interface PageInfo {
  hasNextPage: boolean; /** When paginating, are there more items? */
}
// ...
export interface previous {
  fields?: fields_2 | null; 
  frontmatter?: frontmatter_2 | null; 
}

export interface fields_2 {
  slug?: string | null; 
}

export interface frontmatter_2 {
  title?: string | null; 
}

export interface next {
  fields?: fields_3 | null; 
  frontmatter?: frontmatter_3 | null; 
}
// ...
export interface MarkdownRemark extends Node {
  id: string; /** The id of this node. */
  parent?: Node | null; /** The parent of this node. */
  children?: (Node | null)[] | null; /** The children of this node. */
  internal?: internal_12 | null; 
  frontmatter?: frontmatter_4 | null; 
  rawMarkdownBody?: string | null; 
  fileAbsolutePath?: string | null; 
  fields?: fields_4 | null; 
  html?: string | null; 
  htmlAst?: JSON | null; 
  excerpt?: string | null; 
  headings?: (MarkdownHeading | null)[] | null; 
  timeToRead?: number | null; 
  tableOfContents?: string | null; 
  wordCount?: wordCount | null; 
}

export interface internal_12 {
  content?: string | null; 
  type?: string | null; 
  contentDigest?: string | null; 
  owner?: string | null; 
  fieldOwners?: fieldOwners_2 | null; 
}

export interface fieldOwners_2 {
  slug?: string | null; 
}

export interface frontmatter_4 {
  title?: string | null; 
  _PARENT?: string | null; 
  createdDate?: Date | null; 
  date?: Date | null; 
  author?: string | null; 
  tags?: (string | null)[] | null; 
  draft?: boolean | null; 
}

export interface fields_4 {
  slug?: string | null; 
}

export interface MarkdownHeading {
  value?: string | null; 
  depth?: number | null; 
}

export interface wordCount {
  paragraphs?: number | null; 
  sentences?: number | null; 
  words?: number | null; 
}
// ...
export interface MarkdownRemarkConnection {
  pageInfo: PageInfo; /** Information to aid in pagination. */
  edges?: (MarkdownRemarkEdge | null)[] | null; /** A list of edges. */
  totalCount?: number | null; 
  distinct?: (string | null)[] | null; 
  group?: (markdownRemarkGroupConnectionConnection | null)[] | null; 
}
/** An edge in a connection. */
export interface MarkdownRemarkEdge {
  node?: MarkdownRemark | null; /** The item at the end of the edge */
  next?: MarkdownRemark | null; /** The next edge in the connection */
  previous?: MarkdownRemark | null; /** The previous edge in the connection */
}
/** A connection to a list of items. */
export interface markdownRemarkGroupConnectionConnection {
  pageInfo: PageInfo; /** Information to aid in pagination. */
  edges?: (markdownRemarkGroupConnectionEdge | null)[] | null; /** A list of edges. */
  field?: string | null; 
  fieldValue?: string | null; 
  totalCount?: number | null; 
}
/** An edge in a connection. */
export interface markdownRemarkGroupConnectionEdge {
  node?: MarkdownRemark | null; /** The item at the end of the edge */
  next?: MarkdownRemark | null; /** The next edge in the connection */
  previous?: MarkdownRemark | null; /** The previous edge in the connection */
}